Feature #43: Organize the M2::AUTOSARTemplates::SWComponentTemplate::Communication
