+++ src/armodel/models/M2/MSR/DataDictionary/DataDefProperties.py
@@ -294,3 +294,17 @@ class SwPointerTargetProps(ARObject):
         self.targetCategory = value
         return self
 
+
+class ValueList(ARObject):
+    def __init__(self):
+        super().__init__()
+
+        self.v = None                                       # type: ARFloat
+        self._vf = []                                       # type: List[ARLiteral]
+
+    def addVf(self, vf: ARLiteral):
+        self._vf.append(vf)
+
+    def getVfs(self) -> List[ARLiteral]:
+        return sorted(self._vf)
+
